\name{hvidovre}
\alias{hvidovre}
\docType{data}
\title{Houseprice data}
\description{A dataset with  trade price and attributes for 900 homes in a Danish munnicipality}
\usage{data("hvidovre")}
\format{  A data frame with 901 observations on the following 19 variables
  \describe{
    \item{\code{lprice}}{a numeric vector, price, log price in 1000 EUR}
    \item{\code{brick}}{a numeric vector, dummy, wall made out of brick =1}
    \item{\code{roof_tile}}{a numeric vector, roof_tile}{dummy, roof made out of tiles =1}
    \item{\code{roof_cemen}}{a numeric vector, dummy, roof made out of cement}
    \item{\code{y7}}{a numeric vector, home sold in 2007}
    \item{\code{y8}}{a numeric vector, home sold in 2008}
    \item{\code{y9}}{a numeric vector, home sold in 2009}
    \item{\code{rebuild70}}{a numeric vector, home rebuild in 1970's}
    \item{\code{rebuild80}}{a numeric vector, home rebuild in 1980's}
    \item{\code{rebuild90}}{a numeric vector, home rebuild in 1990's}
    \item{\code{rebuild00}}{a numeric vector,home rebuild in 2000's}
    \item{\code{area}}{a numeric vector,living area in square meters}
    \item{\code{age}}{a numeric vector, build year}
    \item{\code{bathrooms}}{a numeric vector, number of bathrooms}
    \item{\code{highways}}{a numeric vector, distance to nearest highway in meters}
    \item{\code{big_roads}}{a numeric vector, distance to nearest large road in meters}
    \item{\code{railways}}{a numeric vector, railways}{distance to nearest railway in meters}
    \item{\code{nature_SLD}}{a numeric vector, nature_SLD}{distance to nearest nature area in meters}
    \item{\code{lake_SLD}}{a numeric vector,distance to nearest lake in meters}
  }
}
\source{Panduro et al (in review JAERE)
}
\references{
Panduro T.E., Jensen, C.U, Lundhede, T.H., von Graevenitz, K. and Thorsen, B.J., Estimating demand schedules in hedonic analysis: The case of urban parks, (in review JAERE)
}
\examples{data(hvidovre)}
\keyword{datasets}






